pesquisa cabeleireiro

+++ projeto-cabeleireiro/src/main/java/br/com/cabeleireiro/repository/RoleRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import br.com.cabeleireiro.domain.Role;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 
 	Role findByNome(String nome);
